update ipex modeling and add case for text-classification
